Refactor entity style provider impl to dedicated class
