+++ frontend/src/stores/colors/EntityStyle.ts
@@ -26,3 +26,6 @@ export interface NodeStyle extends EntityStyle {
    */
   stroke: NodeStrokeStyle;
 }
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface EdgeStyle extends EntityStyle {}
